Some initial cleanup and re-arranging to assist with addressing issue #64
