+++ .swiftlint.yml
@@ -6,6 +6,7 @@ disabled_rules: # rule identifiers turned on by default to exclude from running
   - large_tuple
   - force_cast
   - type_body_length
+  - trailing_whitespace
 
 excluded: # paths to ignore during linting. Takes precedence over `included`.
   - ./build/SourcePackages/checkouts/swift-argument-parser/*
